refactor(mcp): route runResolvedCommand through invocation kernel
Split server.ts (1061 LOC) along responsibility seams so MCP routes
through the kernel without tripping the complexity gate:

- src/mcp/dispatch.ts      — kernel-backed tool-call dispatcher.
  Hardening errors surface as { isError: true,
  structuredContent.data.suggestion }; warnings flow to
  _meta.warnings. Flat `params` shape preserved (D2 compat).
- src/mcp/tools.ts         — default/expanded/deferred tool builders
  + expandedRegistry.
- src/mcp/handler.ts       — JSON-RPC method dispatch + meta-tools.
- src/mcp/http-transport.ts — POST /mcp simple transport.
- src/mcp/server.ts        — thin bootstrap: load → tools → wire →
  serve.

Only injects default `limit: 20` when the adapter declares a `limit`
arg — ajv strict mode rejects undeclared properties.
